+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added `Explorer.Series.rle_id/1`, which assigns a run-length encoding (RLE) ID
+  to each row, incrementing every time the value changes from one row to the next.
+
 ## [v0.12.0] - 2026-07-05
 
 ### Updated